Centralize markdown rendering
